Add to cart
Add to cart function is complete. However when adding to cart the item should only be allowed to be added to the cart only once. This causes a problem for us. The next methods to be implemented are the delete and the  add to wishlist or save for later methods